Extract Extension trait
