+++ Cargo.toml
@@ -27,6 +27,7 @@ members = [
   "oak_functions/examples/weather_lookup/module",
   "oak_functions/experimental/tf_inference",
   "oak_functions/experimental/metrics",
+  "oak_functions/extension",
   "oak_functions/load_test",
   "oak_functions/loader",
   "oak_functions/location_utils",
