+++ oak_functions/extension/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "oak_functions_extension"
+version = "0.1.0"
+authors = ["Maria Schett <mschett@google.com>"]
+edition = "2021"
+license = "Apache-2.0"
+
+[dependencies]
+anyhow = "*"
+# Use wasmi in `no_std` mode.
+oak_functions_abi = { path = "../abi" }
+wasmi = { version = "*", default-features = false, features = ["core"] }
