+++ oak_functions/lookup/Cargo.toml
@@ -9,9 +9,14 @@ license = "Apache-2.0"
 default = []
 
 [dependencies]
+anyhow = "*"
 log = "*"
 oak_logger = { path = "../logger" }
 oak_functions_util = { path = "../util" }
+oak_functions_abi = { path = "../abi" }
+oak_functions_extension = { path = "../extension" }
+# Use wasmi in `no_std` mode.
+wasmi = { version = "*", default-features = false, features = ["core"] }
 
 [dev-dependencies]
 maplit = "*"
